backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
langchain==0.1.0
langchain-aws==0.1.0
langchain-pinecone==0.2.13
pinecone==5.0.1
boto3==1.34.0
pypdf==3.17.4
pandas>=1.5.0,<2.0.0
openpyxl==3.1.2
python-docx==1.1.0
python-dotenv==1.0.0
unstructured[local-inference]==0.10.30